refactor(tangle-dapp): Move things around a bit
