+++ apps/tangle-dapp/src/components/UnbondingStatsItem.tsx
@@ -4,11 +4,11 @@ import { EMPTY_VALUE_PLACEHOLDER } from '@webb-tools/webb-ui-components/constant
 import addCommasToNumber from '@webb-tools/webb-ui-components/utils/addCommasToNumber';
 import { type FC, useMemo } from 'react';
 
-import useActiveAccountAddress from '../hooks/useActiveAccountAddress';
 import useUnbondingAmount from '../data/nomination/useUnbondingAmount';
 import useUnbonding from '../data/staking/useUnbonding';
 import formatTangleBalance from '../utils/formatTangleBalance';
 import { NominatorStatsItem } from './NominatorStatsItem';
+import useActiveAccountAddress from '@webb-tools/tangle-shared-ui/hooks/useActiveAccountAddress';
 
 const UnbondingStatsItem: FC = () => {
   const activeAccountAddress = useActiveAccountAddress();
