+++ apps/tangle-dapp/src/data/claims/useAirdropEligibility.ts
@@ -4,7 +4,7 @@ import { isEthereumAddress } from '@polkadot/util-crypto';
 import useApiRx from '@webb-tools/tangle-shared-ui/hooks/useApiRx';
 import { useCallback, useEffect, useState } from 'react';
 
-import useActiveAccountAddress from '../../hooks/useActiveAccountAddress';
+import useActiveAccountAddress from '@webb-tools/tangle-shared-ui/hooks/useActiveAccountAddress';
 
 const useAirdropEligibility = () => {
   const [isEligible, setIsEligible] = useState<boolean | null>(null);
