+++ libs/tangle-shared-ui/src/hooks/useViemPublicClient.ts
@@ -1,8 +1,8 @@
-import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
 import { useEffect, useState } from 'react';
 import { createPublicClient, http, type PublicClient } from 'viem';
 
-import createTangleViemChainFromNetwork from '../utils/evm/createTangleViemChainFromNetwork';
+import createTangleViemChainFromNetwork from '../utils/createTangleViemChainFromNetwork';
+import useNetworkStore from '../context/useNetworkStore';
 
 const useViemPublicClient = () => {
   const [publicClient, setPublicClient] = useState<PublicClient | null>(null);
