+++ libs/tangle-shared-ui/src/queries/restake/assetDetails.ts
@@ -7,7 +7,6 @@ import type {
 } from '@polkadot/types/lookup';
 import { BN, formatBalance, hexToString } from '@polkadot/util';
 import type { Chain } from '@webb-tools/dapp-config';
-import { assertEvmAddress } from '@webb-tools/webb-ui-components';
 import { isEvmAddress } from '@webb-tools/webb-ui-components/utils/isEvmAddress20';
 import { combineLatest, map, Observable, of, switchMap } from 'rxjs';
 import { RestakeAssetId } from '../../types';
@@ -16,6 +15,7 @@ import assertRestakeAssetId from '../../utils/assertRestakeAssetId';
 import createAssetIdEnum from '../../utils/createAssetIdEnum';
 import { fetchTokenPriceBySymbol } from '../../utils/fetchTokenPrices';
 import filterNativeAsset from '../../utils/restake/filterNativeAsset';
+import { findErc20Token } from '../../hooks/useTangleEvmErc20Balances';
 
 function createVaultId(u32: Option<u32>): number | null {
   if (u32.isNone) {
@@ -89,14 +89,7 @@ function processAssetDetailsRx(
             const price = null;
 
             if (isEvmAddress(assetId)) {
-              const erc20Token = {
-                name: "Yuri's Local ERC-2 Dummy",
-                symbol: 'USDC',
-                decimals: 18,
-                contractAddress: assertEvmAddress(
-                  '0x2af9b184d0d42cd8d3c4fd0c953a06b6838c9357',
-                ),
-              };
+              const erc20Token = findErc20Token(assetId);
 
               if (erc20Token === null) {
                 return assetMap;
