+++ libs/tangle-shared-ui/src/utils/convertDecimalToBn.ts
