+++ libs/tangle-shared-ui/src/utils/fetchErc20TokenBalance.ts
